+++ CefSharp/IResourceHandlerResponse.cs
@@ -37,7 +37,7 @@ namespace CefSharp
         /// and causes CefSharp to read the response stream in pieces. Thus, setting a length
         /// is optional but allows for more optimal response reading.
         /// </summary>
-        int ContentLength { get; set; }
+        long ContentLength { get; set; }
 
         /// <summary>
         /// URL to redirect to (leave empty to not redirect).
